move most of test helpers into their own crate
