+++ axum-on-rails/Cargo.toml
@@ -1,12 +1,10 @@
 [package]
-name = "rust_rest"
+name = "axum-on-rails"
 version = "0.1.0"
 edition = "2021"
-default-run = "rust_rest"
 
 [dependencies]
 axum = "0.6"
-axum-on-rails = { path = "./axum-on-rails" }
 bb8 = "0.8"
 bb8-postgres = "0.8"
 clap = "4.4"
@@ -16,6 +14,7 @@ refinery = { version = "0.8", features = ["tokio-postgres"]}
 serde = { version = "1.0", features = ["derive"] } 
 tokio = { version = "1.33", features = ["full"] }
 tokio-postgres = "0.7"
+tower = { version = "0.4", features = ["util"] }
 tower-http = { version = "0.4", features = ["full"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
@@ -23,6 +22,4 @@ validator = { version = "0.16", features = ["derive"] }
 
 [dev-dependencies]
 serde_json = "1.0"
-tower = { version = "0.4", features = ["util"] }
 hyper = { version = "0.14", features = ["full"] }
-axum-on-rails-procs = { path = "./axum-on-rails-procs" }
