+++ axum-on-rails/src/lib.rs
@@ -1,0 +1,7 @@
+use bb8::Pool;
+use bb8_postgres::PostgresConnectionManager;
+use tokio_postgres::NoTls;
+
+pub mod test_helpers;
+
+pub type ConnectionPool = Pool<PostgresConnectionManager<NoTls>>;
